// src/services/rest-api/customer-service/model/bank-info-model.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Customer Service API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.267
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface BankInfoModel
 */
export interface BankInfoModel {
    /**
     * 
     * @type {string}
     * @memberof BankInfoModel
     */
    'bankCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof BankInfoModel
     */
    'bank'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof BankInfoModel
     */
    'accountNo'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof BankInfoModel
     */
    'accountName'?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof BankInfoModel
     */
    'isDefault'?: boolean;
}